+++ CHANGELOG.md
@@ -35,6 +35,10 @@
   [Steven Grosmark](https://github.com/g-mark)
   [#3654](https://github.com/realm/SwiftLint/pull/3654)
 
+* Add `anonymous_argument_in_multiline_closure` opt-in rule to validate that
+  named arguments are used in closures that span multiple lines.  
+  [Marcelo Fabri](https://github.com/marcelofabri)
+
 #### Bug Fixes
 
 * Fix false positives in `empty_enum_arguments` rule when comparing values